refactor: divide into modules (analyzer, matcher, template)
